Refactor of facilities synchronization service
